requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
opencv-python>=4.6.0
Pillow>=9.0.0
torch>=2.0.0
torchvision>=0.15.0
requests>=2.28.0
numpy<2
supervision>=0.22.0
